[feat] postId 타입 변경
1. 기존 postId의 타입이 long으로 변경되어 이에 맞추어 변경

+++ src/main/java/me/sisyphusj/community/app/post/mapper/PostMapper.java
@@ -19,7 +19,7 @@ public interface PostMapper {
 
 	int selectTotalCount();
 
-	Optional<PostDetailVO> selectPostDetails(int postId);
+	Optional<PostDetailVO> selectPostDetails(long postId);
 
-	int updateViewsAndGet(int postId);
+	int updateViewsAndGet(long postId);
 }
